pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "An AI-powered coding agent built with Python"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "python-dotenv",
    "pydantic",
    "pyautogen",
    "volcengine-python-sdk",
    "rich",
]

[tool.setuptools.packages.find]
where = ["src"]
